Scroll grid under mouse cursor Move files directly to recycle bin without parent "FFS 2012-05-15 131513" temporary folders Offer $HOME directory alias in directory drop down list (Linux) Support for tilde (~) character in input folder paths (Linux) New environment variables for RealTimeSync: %change_action%, "%change_path% Use Internet Explorer proxy settings for new version check (Windows) Show proper error message after failed symlink creation Start comparison upon double-clicking config list New batch return code: "Synchronization completed with warnings" Hide files that won't be copied by default if direction "none" is part of the rule set (e.g. update variant) Remember save config and folder picker dialog positions separately New sync completion sound Fixed sync completion sound not playing (Ubuntu)

+++ lib/db_file.cpp
@@ -299,8 +299,8 @@ public:
             BinStreamIn& in1stPart = has1stPartL ? inL : inR;
             BinStreamIn& in2ndPart = has1stPartL ? inR : inL;
 
-            const size_t size1stPart = readNumber<std::uint64_t>(in1stPart);
-            const size_t size2ndPart = readNumber<std::uint64_t>(in2ndPart);
+            const size_t size1stPart = static_cast<size_t>(readNumber<std::uint64_t>(in1stPart));
+            const size_t size2ndPart = static_cast<size_t>(readNumber<std::uint64_t>(in2ndPart));
 
             BinaryStream tmpB;
             tmpB.resize(size1stPart + size2ndPart);
